requirements.txt
Appium-Python-Client==2.8.1
behave==1.2.6
selenium==4.8.3

